-start state refactor
